fix fill in blank input text alignment (#11979)
* wip

* fix test

* just update how we write the tests

* update final test

* reorganize code
